+++ TShockAPI/Bouncer.cs
@@ -365,7 +365,7 @@ namespace TShockAPI
 						args.Handled = true;
 						return;
 					}
-					if (action == EditAction.PlaceTile && (editData == TileID.PiggyBank || editData == TileID.Safes) && Main.ServerSideCharacter)
+					if (action == EditAction.PlaceTile && personalStorageTileIDs.Contains(editData) && Main.ServerSideCharacter)
 					{
 						TShock.Log.ConsoleDebug("Bouncer / OnTileEdit rejected from (sscprotect) {0} {1} {2}", args.Player.Name, action, editData);
 						args.Player.SendErrorMessage("You cannot place this tile because server side characters are enabled.");
@@ -2133,6 +2133,14 @@ namespace TShockAPI
 			TileID.LunarMonolith,
 			TileID.TargetDummy,
 			TileID.Campfire
+		};
+
+		private static List<int> personalStorageTileIDs = new List<int>()
+		{
+			TileID.PiggyBank,
+			TileID.Safes,
+			TileID.DefendersForge,
+			TileID.VoidVault
 		};
 
 		/// <summary>